update(bcrypt): version 4 update
- v3 compatible definition created
- minor changes to method definitions to reflect sync/asynchronous with
callbacks/asynchronous wiht Promise support
- JSDocs udpated
- minor typos fixed
- tests rewritten to cover:
  - sync
  - async with callback
  - async with Promise
  - async/await
usage.
- maintainer added

https://github.com/kelektiv/node.bcrypt.js#usage
https://github.com/kelektiv/node.bcrypt.js/compare/v3.0.8...v4.0.1

Thanks!
