+++ types/bcrypt/bcrypt-tests.ts
@@ -1,65 +1,34 @@
+import bcrypt = require('bcrypt');
 
-import bcrypt = require("bcrypt");
+const saltRounds = 10;
+const myPlaintextPassword = 's0//P4$$w0rD';
+const hash = 'fake-hash';
 
-let num: number;
-let str: string;
-let bool: boolean;
-let error: Error;
-
-str = bcrypt.genSaltSync();
-str = bcrypt.genSaltSync(num);
-str = bcrypt.genSaltSync(num, str);
-
-bcrypt.genSalt(num, (err: Error, salt: string): void => { error = err; str = salt; })
-      .then(salt => str = salt)
-      .catch(err => error = err);
-
-bcrypt.genSalt(num, str, (err: Error, salt: string): void => { error = err; str = salt; })
-      .then(salt => str = salt)
-      .catch(err => error = err);
-
-bcrypt.genSalt(num)
-      .then(salt => str = salt)
-      .catch(err => error = err);
-
-bcrypt.genSalt(num, str)
-      .then(salt => str = salt)
-      .catch(err => error = err);
-
-bcrypt.genSalt((err: Error, salt: string): void => { error = err; str = salt; })
-      .then(salt => str = salt)
-      .catch(err => error = err);
-
-bcrypt.genSalt()
-      .then(salt => str = salt)
-      .catch(err => error = err);
-
-
-str = bcrypt.hashSync(str, str);
-str = bcrypt.hashSync(str, num);
-
-bcrypt.hash(str, str, (err: Error, encrypted: string):void => { str = encrypted; })
-      .then(encrypted => str = encrypted)
-      .catch(err => error = err);
-bcrypt.hash(str, str)
-      .then(encrypted => str = encrypted)
-      .catch(err => error = err);
-
-bcrypt.hash(str, num, (err: Error, encrypted: string): void => { str = encrypted; })
-      .then(encrypted => str = encrypted)
-      .catch(err => error = err);
-bcrypt.hash(str, num)
-      .then(encrypted => str = encrypted)
-      .catch(err => error = err);
-
-
-bool = bcrypt.compareSync(str, str);
-
-bcrypt.compare(str, str, (err: Error, same: boolean): void => { bool = same; })
-      .then(same => bool = same)
-      .catch(err => error = err);
-bcrypt.compare(str, str)
-      .then(same => bool = same)
-      .catch(err => error = err);
-
-num = bcrypt.getRounds(str);
+// $ExpectType void
+bcrypt.genSalt(saltRounds, 'a', (err, salt): void => {
+    err; // $ExpectType Error
+    salt; // $ExpectType string
+});
+bcrypt.genSalt(saltRounds, 'b'); // $ExpectType Promise<string>
+// $ExpectType string
+bcrypt.genSaltSync(saltRounds);
+// $ExpectType void
+bcrypt.hash(myPlaintextPassword, saltRounds, (err, hash) => {
+    err; // $ExpectType Error
+    hash; // $ExpectType string
+});
+bcrypt.hash(myPlaintextPassword, saltRounds); // $ExpectType Promise<string>
+bcrypt.hashSync(myPlaintextPassword, saltRounds); // $ExpectType string
+// $ExpectType void
+bcrypt.compare(myPlaintextPassword, hash, (err, result) => {
+    err; // $ExpectType Error
+    result; // $ExpectType boolean
+});
+bcrypt.compare(myPlaintextPassword, hash); // $ExpectType Promise<boolean>
+bcrypt.compareSync(myPlaintextPassword, hash); // $ExpectType boolean
+// async/await https://github.com/kelektiv/node.bcrypt.js#with-promises
+(async () => {
+    await bcrypt.genSalt(saltRounds, 'b'); // $ExpectType string
+    await bcrypt.hash(myPlaintextPassword, saltRounds); // $ExpectType string
+    await bcrypt.compare(myPlaintextPassword, hash); // $ExpectType boolean
+})();
