+++ types/bcrypt/tslint.json
@@ -1,13 +1,1 @@
-{
-    "extends": "dtslint/dt.json",
-    "rules": {
-        "callable-types": false,
-        "no-consecutive-blank-lines": false,
-        "no-redundant-jsdoc": false,
-        "prefer-method-signature": false,
-        "strict-export-declare-modifiers": false,
-        "trim-file": false,
-        "typedef-whitespace": false,
-        "whitespace": false
-    }
-}
+{ "extends": "dtslint/dt.json" }
